Address PR #651 review comments (cvolk, steple, greptile)
- cvolk: move ``StaticPickAndPlaceMimicEnvCfg`` + the ``_no_op_subtask_config``
  helper into ``isaaclab_arena/tasks/pick_and_place_task.py`` next to the
  parent ``G1PickAndPlaceMimicEnvCfg``, and delete the now-empty
  ``isaaclab_arena/tasks/static_pick_and_place_task.py``. Updates the env,
  test, and step_1 doc dropdown to import from the new location.

- greptile P1: drop the unused ``extra_channels`` second positional from
  ``_build_static_mimic_cfg`` in the env, test, and step_1 doc dropdown.
  ``PickAndPlaceTask.get_mimic_env_cfg`` calls the factory with one arg
  (``arm_mode``); the leftover second-arg signature would have raised
  ``TypeError`` at any datagen-time call.

- steple: docs polish across ``static_apple/*.rst``:
  - ``no-navigation``/``no nav`` -> ``no-locomotion`` (reads better and
    matches the rest of the page).
  - ``G1`` -> ``Unitree G1`` for prose mentions of the robot (page title,
    task description, sibling-doc link, step_2 intro, step_4 intro).
  - ``isaaclab docker container`` -> ``Isaac Lab Docker container``.
  - ``loco-manip`` -> ``loco-manipulation`` everywhere.
  - step_2: explicitly mark the ``ufw`` firewall example as Ubuntu-only and
    name the equivalents on other distros.
  - step_3: drop the ``N1.6/N1.7-agnostic`` framing in the LeRobot intro;
    the converter is GR00T-version-agnostic full stop.
  - step_3 + step_4: hoist the multi-line explanatory ``#`` comments out of
    code blocks into surrounding prose. Short single-line YAML section
    markers are kept inside code blocks where they help orient the reader.

No public API changes.
